+++ Calculator.Lib/Interfaces/IOperationFactory.cs
@@ -1,4 +1,4 @@
-namespace Calculator.Lib;
+namespace Calculator.Lib.Interfaces;
 
 public interface IOperationFactory
 {
